# swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/lamp/{deviceId}/bulb-on:
    put:
      tags:
      - lamp-controller
      operationId: turnBulbOn
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/lamp/{deviceId}/bulb-off:
    put:
      tags:
      - lamp-controller
      operationId: turnBulbOff
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/lamp/{deviceId}/automatic:
    put:
      tags:
      - lamp-controller
      operationId: setAutomatic
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: val
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /api/images/property/upload:
    post:
      tags:
      - image-controller
      operationId: uploadPropertyImage
      parameters:
      - name: propertyId
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMessageDTO'
  /api/images/device/upload:
    post:
      tags:
      - image-controller
      operationId: uploadDeviceImage
      parameters:
      - name: customFileName
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMessageDTO'
...
  /api/lamp/{deviceId}:
    get:
      tags:
      - lamp-controller
      operationId: getLamp
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDTO'
  /api/lamp/{deviceId}/range:
    get:
      tags:
      - lamp-controller
      operationId: getRangeLightSensor
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: start
        in: query
        required: true
        schema:
          type: string
      - name: end
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListValueTimestampDTO'
  /api/lamp/{deviceId}/range/bulb:
    get:
      tags:
      - lamp-controller
      operationId: getDateRangeBulb
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: start
        in: query
        required: true
        schema:
          type: string
      - name: end
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListBulbOnOffMeasurement'
...
components:
  schemas:
...
    DeviceDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        state:
          type: boolean
        modelName:
          type: string
        usesElectricity:
          type: boolean
        consumptionAmount:
          type: number
          format: double
        propertyId:
          type: integer
          format: int32
        imagePath:
          type: string
        deviceType:
          type: string
        "on":
          type: boolean
...
    MessageDTO:
      required:
      - message
      - topic
      type: object
      properties:
        message:
          type: string
        topic:
          type: string
    ApiResponseMessageDTO:
      type: object
      properties:
        status:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/MessageDTO'
...
    ApiResponseListValueTimestampDTO:
      type: object
      properties:
        status:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/ValueTimestampDTO'
    ValueTimestampDTO:
      type: object
      properties:
        value:
          type: string
        timestamp:
          type: string
    ApiResponseListBulbOnOffMeasurement:
      type: object
      properties:
        status:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/BulbOnOffMeasurement'
    BulbOnOffMeasurement:
      type: object
      properties:
        value:
          type: string
        timestamp:
          type: string
...